model_training.py
import tensorflow as tf
import numpy as np
import os
import cv2
from sklearn.model_selection import train_test_split

def create_sample_model():
    """
    Create a simple CNN model for liver tumor detection.
    This is a placeholder function that would be replaced with actual model training
    using a real dataset in a production environment.
    """
    # Create a simple CNN model
    model = tf.keras.Sequential([
        tf.keras.layers.Conv2D(32, (3, 3), activation='relu', input_shape=(224, 224, 3)),
        tf.keras.layers.MaxPooling2D((2, 2)),
        tf.keras.layers.Conv2D(64, (3, 3), activation='relu'),
        tf.keras.layers.MaxPooling2D((2, 2)),
        tf.keras.layers.Conv2D(128, (3, 3), activation='relu'),
        tf.keras.layers.MaxPooling2D((2, 2)),
        tf.keras.layers.Flatten(),
        tf.keras.layers.Dense(128, activation='relu'),
        tf.keras.layers.Dropout(0.5),
        tf.keras.layers.Dense(1, activation='sigmoid')
    ])
    
    # Compile the model
    model.compile(
        optimizer='adam',
        loss='binary_crossentropy',
        metrics=['accuracy']
    )
    
    return model

def generate_dummy_data(num_samples=100):
    """
    Generate dummy data for demonstration purposes.
    In a real scenario, this would be replaced with actual CT scan data.
    """
    # Create random images (224x224x3)
    X = np.random.rand(num_samples, 224, 224, 3)
    
    # Create random labels (0 or 1)
    y = np.random.randint(0, 2, size=(num_samples, 1))
    
    return train_test_split(X, y, test_size=0.2, random_state=42)

def train_model(model, X_train, y_train, X_val, y_val, epochs=5):
    """
    Train the model with the provided data.
    """
    history = model.fit(
        X_train, y_train,
        epochs=epochs,
        batch_size=32,
        validation_data=(X_val, y_val),
        verbose=1
    )
    
    return history

def convert_to_tflite(model, output_path):
    """
    Convert the trained model to TensorFlow Lite format.
    """
    # Create models directory if it doesn't exist
    os.makedirs(os.path.dirname(output_path), exist_ok=True)
    
    # Convert the model to TensorFlow Lite format
    converter = tf.lite.TFLiteConverter.from_keras_model(model)
    tflite_model = converter.convert()
    
    # Save the TFLite model
    with open(output_path, 'wb') as f:
        f.write(tflite_model)
    
    print(f"Model saved to {output_path}")

if __name__ == "__main__":
    # Create the model
    model = create_sample_model()
    
    # Generate dummy data
    X_train, X_val, y_train, y_val = generate_dummy_data()
    
    # Train the model
    history = train_model(model, X_train, y_train, X_val, y_val)
    
    # Convert and save the model
    convert_to_tflite(model, os.path.join('models', 'tumor_model.tflite'))
    
    print("Model training and conversion completed successfully!")