requirements.txt
flask==2.0.1
tensorflow==2.8.0
fastai
numpy>=1.21.2
opencv-python==4.7.0.72
pillow==8.3.1
scikit-image==0.18.3
matplotlib==3.4.3
gunicorn==20.1.0